Add map. Update main list item UI.

+++ app/src/main/java/com/humber/its2020/ibourit/entity/Article.kt
@@ -1,8 +1,8 @@
 package com.humber.its2020.ibourit.entity
 
 data class Article (
-    val id: Int = 0,
-    val userId: Int,
+    val id: Long = 0,
+    val userId: Long,
     val content: String,
     val like: Int,
     val comments: List<String>
