+++ app/src/main/java/com/humber/its2020/ibourit/entity/User.kt
@@ -1,7 +1,7 @@
 package com.humber.its2020.ibourit.entity
 
-data class User(
-    val id: Int,
+data class User (
+    val id: Long,
     val email: String,
     val name: String
 )
